pyproject.toml
[project]
name = "yoonagain-wiki"
version = "0.1.0"
description = "윤어게인 위키 빌더"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "firebase-admin>=6.5.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
compile = "scripts.compile:main"
export = "scripts.export_firestore:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
